feat(user-list): accommodate guest management with user list
