+++ bigbluebutton-html5/imports/ui/components/sidebar-content/component.tsx
@@ -3,10 +3,9 @@ import Resizable from 're-resizable';
 import { ACTIONS, PANELS } from '../layout/enums';
 import ChatContainer from '/imports/ui/components/chat/chat-graphql/component';
 import ProfileSettings from '/imports/ui/components/profile-settings/component';
-import GuestManagement from '../user-list/guest-management/component';
 import NotesContainer from '/imports/ui/components/notes/component';
 import PollContainer from '/imports/ui/components/poll/container';
-import UserListParticipants from '/imports/ui/components/user-list/user-list-participants/component';
+import UserListComponent from '/imports/ui/components/user-list/component';
 import BreakoutRoomContainer from '../breakout-room/breakout-room/component';
 import TimerContainer from '/imports/ui/components/timer/panel/component';
 import GuestUsersManagementPanel from '/imports/ui/components/waiting-users/waiting-users-graphql/component';
@@ -141,9 +140,8 @@ const SidebarContent = (props: SidebarContentProps) => {
             isToSharedNotesBeShow={sidebarContentPanel === PANELS.SHARED_NOTES}
           />
         )}
-        {sidebarContentPanel === PANELS.USERLIST && <GuestManagement />}
         {sidebarContentPanel === PANELS.PROFILE && <ProfileSettings />}
-        {sidebarContentPanel === PANELS.USERLIST && <UserListParticipants />}
+        {sidebarContentPanel === PANELS.USERLIST && <UserListComponent />}
         {sidebarContentPanel === PANELS.BREAKOUT && <BreakoutRoomContainer />}
         {sidebarContentPanel === PANELS.TIMER && <TimerContainer />}
         {sidebarContentPanel === PANELS.WAITING_USERS && <GuestUsersManagementPanel />}
