--- bigbluebutton-html5/imports/ui/components/user-list/component.jsx
@@ -1,15 +1,0 @@
-import React, { PureComponent } from 'react';
-import Styled from './styles';
-import UserContentContainer from './user-list-content/container';
-
-class UserList extends PureComponent {
-  render() {
-    return (
-      <Styled.UserList>
-        <UserContentContainer />
-      </Styled.UserList>
-    );
-  }
-}
-
-export default UserList;
