--- bigbluebutton-html5/imports/ui/components/user-list/container.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import UserList from './component';
-
-const UserListContainer = (props) => (
-  <UserList
-    {...props}
-  />
-);
-
-export default UserListContainer;
