+++ bigbluebutton-html5/imports/ui/components/user-list/types.ts
@@ -1,0 +1,1 @@
+export interface UserListComponentProps {}
